+++ .env.sample
@@ -1,18 +1,14 @@
+#######################################################
 # Sample development values 
-#
-# Sxample cluster base domain used: dev.gke.otomi.cloud
-#
 # Tasks will complain when a value is missing ;)
 #######################################################
-
-# DEBUG="*"
 
 # Harbor task
 HARBOR_USER='admin'
 HARBOR_PASSWORD=''
 HARBOR_BASE_URL='http://127.0.0.1:8083/api/v2.0'
 OIDC_CLIENT_SECRET=''
-OIDC_ENDPOINT='https://keycloak.dev.eks.otomi.cloud'
+OIDC_ENDPOINT='https://keycloak.dev.gke.otomi.cloud'
 OIDC_VERIFY_CERT='true' # set to false when staging certs are used on the cluster
 TEAM_NAMES='["team-otomi","team-chai","team-dev","team-demo"]'
 
@@ -24,9 +20,9 @@ IDP_ALIAS='redkubes-azure'
 IDP_GROUP_OTOMI_ADMIN='e69ded30-0882-4490-8e0f-2e67625a0693'
 IDP_GROUP_TEAM_ADMIN='3c63814c-59df-46c3-9a69-d9e1c3611097'
 IDP_GROUP_MAPPINGS_TEAMS='{"team-demo":"28010af7-9535-4265-8689-50f51f8f2c87","team-dev":"3d2e6df8-c7c7-4d51-aa6a-9b7b5330d915","team-otomi":"0efd2f6d-fb8b-49a9-9507-54cd6e92c348"}'
-IDP_USERNAME_CLAIM_MAPPER="${CLAIM.email}"
+IDP_USERNAME_CLAIM_MAPPER='${CLAIM.email}'
 IDP_SUB_CLAIM_MAPPER='oid'
-KEYCLOAK_ADDRESS='https://keycloak.dev.eks.otomi.cloud'
+KEYCLOAK_ADDRESS='https://keycloak.dev.gke.otomi.cloud'
 KEYCLOAK_ADMIN='admin'
 KEYCLOAK_ADMIN_PASSWORD=''
 KEYCLOAK_REALM='master'
@@ -35,8 +31,9 @@ IDP_OIDC_URL='https://login.microsoftonline.com/$TENANT_ID'
 REDIRECT_URIS='["http://localhost:8084/*","https://auth.dev.gke.otomi.cloud/*","https://otomi.dev.gke.otomi.cloud/*","https://harbor.dev.gke.otomi.cloud/*","https://apps.dev.gke.otomi.cloud/*","https://apps.team-dev.dev.gke.otomi.cloud/*","https://apps.team-demo.dev.gke.otomi.cloud/*","https://apps.team-admin.dev.gke.otomi.cloud/*","https://apps.team-otomi.dev.gke.otomi.cloud/*","https://apps.team-chai.dev.gke.otomi.cloud/*"]'
 
 # Gitea task
-GITEA_REPO='values'
-GITEA_USER='otomi'
 GITEA_PASSWORD=''
 GITEA_URL='http://127.0.0.1:8082'
-DRONE_URL='http://drone.dev.eks.otomi.cloud'
+DRONE_URL='https://drone.dev.gke.otomi.cloud'
+
+# Drone task
+DRONE_TOKEN=''
