+++ README.md
@@ -28,6 +28,21 @@ npm run task:(gitea*|harbor|keycloak|certs-aws|...)-dev
 
 Or you can start them in the vscode debugger.
 
+**Skipping TLS (like for staging certs):**
+
+Run the next line in your shell to skip TLS cert validation, like when using self-signed certs like letsencrypt staging:
+```bash
+export NODE_TLS_REJECT_UNAUTHORIZED='0'
+```
+
+**Setting debug level:**
+
+For all packages to turn on debug:
+```bash
+export DEBUG='*'
+```
+To limit scope please read [the debug docs](https://github.com/visionmedia/debug).
+
 ## Unit tests
 
 There are not many unit tests, as the tasks are *very* robust and idempotent. You can run them as always with:
