+++ src/tasks/gitea/common.ts
@@ -1,7 +1,4 @@
+/* eslint-disable import/prefer-default-export */
 import { CustomError } from 'ts-custom-error'
 
 export class GiteaDroneError extends CustomError {}
-export const orgName = 'otomi'
-export const teamName = 'otomi-viewer'
-export const repoName = 'values'
-export const username = 'otomi-admin'
