+++ src/tasks/gitea/gitea-add-users-to-team.ts
@@ -1,22 +1,7 @@
-import {
-  OrganizationApi,
-  // CreateRepoOption,
-  // CreateOrgOption,
-  // CreateTeamOption,
-  AdminApi,
-  User,
-  Team,
-} from '@redkubes/gitea-client-node'
+import { OrganizationApi, AdminApi, User, Team } from '@redkubes/gitea-client-node'
 import { doApiCall } from '../../utils'
 import { cleanEnv, GITEA_PASSWORD, GITEA_URL } from '../../validators'
-import {
-  orgName,
-  teamName,
-  // orgName,
-  // teamName,
-  // repoName,
-  username,
-} from './common'
+import { orgName, teamNameOwners, teamNameViewer, username } from '../common'
 
 const env = cleanEnv({
   GITEA_PASSWORD,
@@ -34,26 +19,31 @@ export default async function main(): Promise<void> {
   const users: User[] = await doApiCall(errors, `Get all users`, () => adminApi.adminGetAllUsers())
 
   const orgApi = new OrganizationApi(username, env.GITEA_PASSWORD, `${giteaUrl}/api/v1`)
-  const teams = (await doApiCall(errors, 'Find team ID if exists', () => orgApi.orgListTeams(orgName))) as Team[]
-  const team = teams.find((team) => team.name === teamName)
-  const teamID = team!.id as number
+  await Promise.allSettled(
+    [teamNameViewer].map(
+      async (teamName): Promise<any[]> => {
+        const orgTeams = (await doApiCall(errors, 'Find team ID if exists', () =>
+          orgApi.orgListTeams(orgName),
+        )) as Team[]
+        const teamID = orgTeams.find((team) => team.name === teamName)!.id as number
 
-  const membersInTeam: User[] = (await doApiCall(errors, 'Find all members in this team', () =>
-    orgApi.orgListTeamMembers(teamID),
-  )) as User[]
-  const userIDsInTeam = membersInTeam.map((member) => member.id)
+        const membersInTeam = (await doApiCall(errors, 'Find all members in this team', () =>
+          orgApi.orgListTeamMembers(teamID),
+        )) as User[]
+        const userIDsInTeam = membersInTeam.map((member) => member.id)
+        const membersNotInTeam: User[] = users.filter((user) => !userIDsInTeam.includes(user.id))
 
-  const membersNotInTeam: User[] = users.filter((user) => !userIDsInTeam.includes(user.id))
-  const promises: Promise<any>[] = []
-  // eslint-disable-next-line no-restricted-syntax
-  for (const member of membersNotInTeam) {
-    promises.push(
-      doApiCall(errors, `Add ${member.login} to ${teamName}`, () =>
-        orgApi.orgAddTeamMember(teamID, member.login as string),
-      ),
-    )
-  }
-  Promise.allSettled(promises)
+        // eslint-disable-next-line no-restricted-syntax
+        return Promise.allSettled(
+          membersNotInTeam.map((member) =>
+            doApiCall(errors, `Add ${member.login} to ${teamName}`, () =>
+              orgApi.orgAddTeamMember(teamID, member.login as string),
+            ),
+          ),
+        )
+      },
+    ),
+  )
 
   if (errors.length) {
     console.error(`Errors found: ${JSON.stringify(errors, null, 2)}`)
